// src/api/model/sensor.ts
/**
 * LinkIOT Dashboard
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Sensor { 
    id: number;
    name: string;
    deviceId: string;
    dataType: number;
    realType: string;
    showType: string;
    description: string;
    updateTime: string;
    createTime: string;
    status: Boolean;
}