// src/api/model/body5.ts
/**
 * LinkIOT Dashboard
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Body5 { 
    /**
     * 传感器ID
     */
    sensorId: number;
    /**
     * 传感器名称
     */
    name?: string;
    /**
     * 数据类型
     */
    dataType?: number;
    /**
     * 数据展示方式
     */
    showType?: string;
    /**
     * 传感器描述
     */
    description?: string;
}